BIP44 support *  Refactor UnlockWalletDialogFragment to use an abstract base  * Rename EncryptNewKeyChainFragment to EncryptNewKeyChainDialogFragment, derived from the abstract base  * Added Upgrade process (if BIP44 is missing, inform the user the wallet will be upgraded)  * Support BIP44 for testnet (m/44'/1'/0) and mainnet (m/44'/5'/0')  * move upgrade method from WalletUtils to WalletActivity  * set version to 5.17.6-lock-bip44
