+++ wallet/AndroidManifest.xml
@@ -3,8 +3,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     package="de.schildbach.wallet_test"
     android:installLocation="internalOnly"
-    android:versionCode="51712"
-    android:versionName="5.17.6" >
+    android:versionCode="51713"
+    android:versionName="5.17.6-lock-bip44" >
 
     <uses-sdk
         android:minSdkVersion="16"
