+++ wallet/res/layout/encrypt_new_key_chain_dialog.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:paddingLeft="@dimen/list_entry_padding_horizontal_lax"
+    android:paddingRight="@dimen/list_entry_padding_horizontal_lax"
+    android:paddingBottom="@dimen/list_entry_padding_vertical"
+    android:divider="@drawable/divider_field"
+    android:orientation="vertical"
+    android:showDividers="middle">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/list_entry_padding_vertical"
+        android:divider="@drawable/divider_field"
+        android:orientation="horizontal"
+        android:showDividers="middle">
+
+        <TextView
+            android:id="@+id/new_key_chain_dialog_message_text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="@dimen/list_entry_padding_vertical"
+            android:layout_marginTop="@dimen/list_entry_padding_vertical"
+            android:layout_weight="2"
+            android:text="@string/encrypt_new_key_chain_dialog_message" />
+
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/list_entry_padding_vertical"
+        android:divider="@drawable/divider_field"
+        android:orientation="horizontal"
+        android:showDividers="middle">
+
+        <EditText
+            android:id="@+id/pin"
+            android:layout_width="0px"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:hint="@string/private_key_password"
+            android:imeOptions="flagNoExtractUi"
+            android:inputType="numberPassword"
+            android:singleLine="true" />
+
+        <Button
+            android:id="@+id/unlock"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_weight="1"
+            android:text="@string/wallet_lock_unlock" />
+
+    </LinearLayout>
+
+    <TextView
+        android:id="@+id/bad_pin"
+        tools:visibility="visible"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/private_key_bad_password"
+        android:textColor="@color/fg_error"
+        android:textStyle="bold"
+        android:visibility="invisible" />
+
+</LinearLayout>
