+++ wallet/res/values/strings-extra.xml
@@ -65,4 +65,10 @@
 		<item quantity="other">hours</item>
 	</plurals>
 
+	<string name="encrypt_new_key_chain_dialog_title">Upgrade your wallet</string>
+	<string name="encrypt_new_key_chain_dialog_message">Your wallet will be upgraded to support importing from and exporting to more apps.</string>
+	<string name="encrypt_new_key_chain_enter_pin_dialog_message">You must enter your PIN to finish the upgrade process.</string>
+	<string name="pin_code_required_dialog_message">This app now requires a PIN to spend and view the balance and transactions.</string>
+
+
 </resources>
